Major update.
Due to being kicked out of floor 8 before accomplishing my issues with
the Spatial Mapping I had to take a different approach. The program is
now setup with world anchors.  What this means is that now, when you
load the program you have two options, edit mode and experience. In edit
mode the objects that need placing will surround the user. They will be
able to click on the objects and move them to their appropriate spots in
the program. The benefit to world anchros is that when you quit and come
back they will be where you left them( unless you recompile and rebuild.